Full Wordpress Layout
 On branch master

	modified:   dev/manage.sh
	new file:   wp-content/themes/lazycrazyacres/content.php
	modified:   wp-content/themes/lazycrazyacres/footer.php
	new file:   wp-content/themes/lazycrazyacres/functions.php
	modified:   wp-content/themes/lazycrazyacres/header.php
	modified:   wp-content/themes/lazycrazyacres/index.php
	copied:     wp-content/themes/lazycrazyacres/index.php -> wp-content/themes/lazycrazyacres/page.php
	modified:   wp-content/themes/lazycrazyacres/sidebar.php
	modified:   wp-content/themes/twentyeleven/index.php

+++ dev/manage.sh
@@ -90,7 +90,9 @@ mirror () {
     --exclude-from=$EXCLUDE_LIST \
     $REPO_DIR/ $target
 
-    cp "$REPO_DIR/htaccess-dev" "$target/.htaccess"
+    if [ -f "$REPO_DIR/htaccess-dev" ]; then
+        cp "$REPO_DIR/htaccess-dev" "$target/.htaccess"
+    fi
 }
 
 main "$@"
